Most of the AVR ADC code moved to adc_driver.cpp
Based on suggestions from Bertrand. Tested only on a modded 9XR128. May need
regression testing on other hardware variants.
