+++ radio/src/targets/gruvin9x/board_gruvin9x.cpp
@@ -65,8 +65,7 @@ inline void boardInit()
   DDRL = 0x80;  PORTL = 0xff; // 7: Hold_PWR_On (1=On, default Off), 6:Jack_Presence_TTL, 5-0: User Button inputs
 #endif
 
-  ADMUX=ADC_VREF_TYPE;
-  ADCSRA=0x85; // ADC enabled, pre-scaler division=32 (no interrupt, no auto-triggering)
+  adcInit();
   ADCSRB=(1<<MUX5);
 
   /**** Set up timer/counter 0 ****/
